requirement.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
opencv-python
numpy
pillow
pandas